FEAT-2898: tun: IFF_RX_CSUM_FIXUP calculates checksum of rx'd packets
If a tun/tap device has IFF_RX_CSUM_FIXUP enabled then the checksum of
received packets destinted for vhost-user/-net (i.e., packets being
sent to a virtio-net frontend in the guest) will have their checksums
computed.

This is supposed to improve performance of Windows virtio frontends
because the checksum recalculation is combined with a copy instead of
the frontend having to recompute it in another loop. However, since
this is always enabled instead of enabled by a feature request from
the frontend driver, it presumably reduces performance of Linux
guests.

[ Based on commit 12d7366b2b75e916268b437e3922720b243d3c48 and
  d0c812de17c26df16fe0329911b18af1af7f70c1 forward ported to 4.14. ]

Change-Id: If1e464c79771eaa00bf22c6c5c66537bec76d407