sysrq w, t: don't output scheduling info
The call to sysrq_sched_debug_show pollutes the output of sysrq keys w
and t with a lot of scheduling information. Often this information is
more than enough to fill the dmesg buffer and is rarely what you want
when triggering those sysrqs, hence it should be disabled. Anyone who
does want this information can still access it via /proc/sched_debug.

Note that we can't simply unset CONFIG_SCHED_DEBUG because that also
controls the existence of /proc/<pid>/sched which is inspected by
acropolis.

Change-Id: Ia5daeee7ffdfbe85720aacbc65815d26d07bfbbb